Restructured the component
